PEAR-2348 - rename test file
